+++ backend/config/jwt.js
@@ -1,0 +1,5 @@
+const JWT_SECRET = 'your_jwt_secret_key'; // Replace with a secure, random key in production
+
+module.exports = {
+  JWT_SECRET
+};
